provide basic iban validation, parsing and construction
Currently implementation is ignorant of country-specific IBAN
rules. That is, it will accept any IBAN matching
[A-Z]{2}[0-9]{2}[A-Z0-9a-z]{0,30}, provided that the checkdigits are
consistent with the content.

This would allow one to generate IBANs with invalid BBAN structure
simply by computing the correct check digits. It also allows one to
generate IBANs with uncheckable BBAN structure by using a undefined
country-code or a country-code for a country that's not yet part of
the IBAN registry.